requirements.txt
fastapi
uvicorn
spacy
pdfminer.six
python-multipart
python-docx
